requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
meilisearch>=0.40.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
agno>=2.0.0
mcp>=1.0.0
